Modularized the Streamlit UI, fixing a couple bugs
